+++ test/cypress/tests/utils/popover.cy.ts
@@ -1,6 +1,7 @@
-import { PopoverDesktop as Popover } from '../../../../src/components/utils/popover';
-import { PopoverItem } from '../../../../types';
-import { TunesMenuConfig } from '../../../../types/tools';
+import { PopoverDesktop as Popover, PopoverItemType } from '../../../../src/components/utils/popover';
+import { PopoverItemParams } from '../../../../types';
+import { MenuConfig } from '../../../../types/tools';
+import Header from '@editorjs/header';
 
 /* eslint-disable @typescript-eslint/no-empty-function */
 
@@ -15,13 +16,13 @@ describe('Popover', () => {
      * Confirmation is moved to separate variable to be able to test it's callback execution.
      * (Inside popover null value is set to confirmation property, so, object becomes unavailable otherwise)
      */
-    const confirmation = {
+    const confirmation: PopoverItemParams = {
       icon: confirmActionIcon,
       title: confirmActionTitle,
       onActivate: cy.stub(),
     };
 
-    const items: PopoverItem[] = [
+    const items: PopoverItemParams[] = [
       {
         icon: actionIcon,
         title: actionTitle,
@@ -69,7 +70,7 @@ describe('Popover', () => {
   });
 
   it('should render the items with true isActive property value as active', () => {
-    const items: PopoverItem[] = [
+    const items = [
       {
         icon: 'Icon',
         title: 'Title',
@@ -93,7 +94,7 @@ describe('Popover', () => {
   });
 
   it('should not execute item\'s onActivate callback if the item is disabled', () => {
-    const items: PopoverItem[] = [
+    const items: PopoverItemParams[] = [
       {
         icon: 'Icon',
         title: 'Title',
@@ -115,6 +116,10 @@ describe('Popover', () => {
         .should('have.class', 'ce-popover-item--disabled')
         .click()
         .then(() => {
+          if (items[0].type !== PopoverItemType.Default) {
+            return;
+          }
+
           // Check onActivate callback has never been called
           expect(items[0].onActivate).to.have.not.been.called;
         });
@@ -122,7 +127,7 @@ describe('Popover', () => {
   });
 
   it('should close once item with closeOnActivate property set to true is activated', () => {
-    const items: PopoverItem[] = [
+    const items = [
       {
         icon: 'Icon',
         title: 'Title',
@@ -149,7 +154,7 @@ describe('Popover', () => {
   });
 
   it('should highlight as active the item with toggle property set to true once activated', () => {
-    const items: PopoverItem[] = [
+    const items = [
       {
         icon: 'Icon',
         title: 'Title',
@@ -173,7 +178,7 @@ describe('Popover', () => {
   });
 
   it('should perform radiobutton-like behavior among the items that have toggle property value set to the same string value', () => {
-    const items: PopoverItem[] = [
+    const items = [
       {
         icon: 'Icon 1',
         title: 'Title 1',
@@ -218,7 +223,7 @@ describe('Popover', () => {
   });
 
   it('should toggle item if it is the only item in toggle group', () => {
-    const items: PopoverItem[] = [
+    const items = [
       {
         icon: 'Icon',
         title: 'Title',
@@ -241,22 +246,149 @@ describe('Popover', () => {
     });
   });
 
-  it('should render custom html content', () => {
-    const customHtml = document.createElement('div');
+  it('should display item with custom html', () => {
+    /**
+     * Block Tune with html as return type of render() method
+     */
+    class TestTune {
+      public static isTune = true;
 
-    customHtml.setAttribute('data-cy-name', 'customContent');
-    customHtml.innerText = 'custom html content';
-    const popover = new Popover({
-      customContent: customHtml,
-      items: [],
+      /** Tune control displayed in block tunes popover */
+      public render(): HTMLElement {
+        const button = document.createElement('button');
+
+        button.classList.add('ce-settings__button');
+        button.innerText = 'Tune';
+
+        return button;
+      }
+    }
+
+    /** Create editor instance */
+    cy.createEditor({
+      tools: {
+        testTool: TestTune,
+      },
+      tunes: [ 'testTool' ],
+      data: {
+        blocks: [
+          {
+            type: 'paragraph',
+            data: {
+              text: 'Hello',
+            },
+          },
+        ],
+      },
     });
 
-    cy.document().then(doc => {
-      doc.body.append(popover.getElement());
+    /** Open block tunes menu */
+    cy.get('[data-cy=editorjs]')
+      .get('.cdx-block')
+      .click();
 
-      /* Check custom content exists in the popover */
-      cy.get('[data-cy-name=customContent]');
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-toolbar__settings-btn')
+      .click();
+
+    /** Check item with custom html content is displayed */
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-popover .ce-popover-item-html')
+      .contains('Tune')
+      .should('be.visible');
+  });
+
+  it('should support flipping between custom content items', () => {
+    /**
+     * Block Tune with html as return type of render() method
+     */
+    class TestTune1 {
+      public static isTune = true;
+
+      /** Tune control displayed in block tunes popover */
+      public render(): HTMLElement {
+        const button = document.createElement('button');
+
+        button.classList.add('ce-settings__button');
+        button.innerText = 'Tune1';
+
+        return button;
+      }
+    }
+
+    /**
+     * Block Tune with html as return type of render() method
+     */
+    class TestTune2 {
+      public static isTune = true;
+
+      /** Tune control displayed in block tunes popover */
+      public render(): HTMLElement {
+        const button = document.createElement('button');
+
+        button.classList.add('ce-settings__button');
+        button.innerText = 'Tune2';
+
+        return button;
+      }
+    }
+
+    /** Create editor instance */
+    cy.createEditor({
+      tools: {
+        testTool1: TestTune1,
+        testTool2: TestTune2,
+      },
+      tunes: ['testTool1', 'testTool2'],
+      data: {
+        blocks: [
+          {
+            type: 'paragraph',
+            data: {
+              text: 'Hello',
+            },
+          },
+        ],
+      },
     });
+
+    /** Open block tunes menu */
+    cy.get('[data-cy=editorjs]')
+      .get('.cdx-block')
+      .click();
+
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-toolbar__settings-btn')
+      .click();
+
+    /** Press Tab */
+    // eslint-disable-next-line cypress/require-data-selectors -- cy.tab() not working here
+    cy.get('body').tab();
+
+    /** Check the first custom html item is focused */
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-popover .ce-popover-item-html .ce-settings__button')
+      .contains('Tune1')
+      .should('have.class', 'ce-popover-item--focused');
+
+    /** Press Tab */
+    // eslint-disable-next-line cypress/require-data-selectors -- cy.tab() not working here
+    cy.get('body').tab();
+
+    /** Check the second custom html item is focused */
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-popover .ce-popover-item-html .ce-settings__button')
+      .contains('Tune2')
+      .should('have.class', 'ce-popover-item--focused');
+
+    /** Press Tab */
+    // eslint-disable-next-line cypress/require-data-selectors -- cy.tab() not working here
+    cy.get('body').tab();
+
+    /** Check that default popover item got focused */
+    cy.get('[data-cy=editorjs]')
+      .get('[data-item-name=move-up]')
+      .should('have.class', 'ce-popover-item--focused');
   });
 
   it('should display nested popover (desktop)', () => {
@@ -265,7 +397,7 @@ describe('Popover', () => {
       public static isTune = true;
 
       /** Tool data displayed in block tunes popover */
-      public render(): TunesMenuConfig {
+      public render(): MenuConfig {
         return  {
           icon: 'Icon',
           title: 'Title',
@@ -277,6 +409,7 @@ describe('Popover', () => {
                 icon: 'Icon',
                 title: 'Title',
                 name: 'nested-test-item',
+                onActivate: (): void => {},
               },
             ],
           },
@@ -334,14 +467,13 @@ describe('Popover', () => {
       .should('be.visible');
   });
 
-
   it('should display children items, back button and item header and correctly switch between parent and child states (mobile)', () => {
     /** Tool class to test how it is displayed inside block tunes popover */
     class TestTune {
       public static isTune = true;
 
       /** Tool data displayed in block tunes popover */
-      public render(): TunesMenuConfig {
+      public render(): MenuConfig {
         return  {
           icon: 'Icon',
           title: 'Tune',
@@ -353,6 +485,7 @@ describe('Popover', () => {
                 icon: 'Icon',
                 title: 'Title',
                 name: 'nested-test-item',
+                onActivate: (): void => {},
               },
             ],
           },
@@ -440,5 +573,476 @@ describe('Popover', () => {
     cy.get('[data-cy=editorjs]')
       .get('.ce-popover-header')
       .should('not.exist');
+  });
+
+
+  it('should display default (non-separator) items without specifying type: default', () => {
+    /** Tool class to test how it is displayed inside block tunes popover */
+    class TestTune {
+      public static isTune = true;
+
+      /** Tool data displayed in block tunes popover */
+      public render(): MenuConfig {
+        return  {
+          onActivate: (): void => {},
+          icon: 'Icon',
+          title: 'Tune',
+          toggle: 'key',
+          name: 'test-item',
+        };
+      }
+    }
+
+    /** Create editor instance */
+    cy.createEditor({
+      tools: {
+        testTool: TestTune,
+      },
+      tunes: [ 'testTool' ],
+      data: {
+        blocks: [
+          {
+            type: 'paragraph',
+            data: {
+              text: 'Hello',
+            },
+          },
+        ],
+      },
+    });
+
+    /** Open block tunes menu */
+    cy.get('[data-cy=editorjs]')
+      .get('.cdx-block')
+      .click();
+
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-toolbar__settings-btn')
+      .click();
+
+    /** Check item displayed */
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-popover__container')
+      .get('[data-item-name="test-item"]')
+      .should('be.visible');
+  });
+
+  it('should display separator', () => {
+    /** Tool class to test how it is displayed inside block tunes popover */
+    class TestTune {
+      public static isTune = true;
+
+      /** Tool data displayed in block tunes popover */
+      public render(): MenuConfig {
+        return  [
+          {
+            onActivate: (): void => {},
+            icon: 'Icon',
+            title: 'Tune',
+            toggle: 'key',
+            name: 'test-item',
+          },
+          {
+            type: PopoverItemType.Separator,
+          },
+        ];
+      }
+    }
+
+
+    /** Create editor instance */
+    cy.createEditor({
+      tools: {
+        testTool: TestTune,
+      },
+      tunes: [ 'testTool' ],
+      data: {
+        blocks: [
+          {
+            type: 'paragraph',
+            data: {
+              text: 'Hello',
+            },
+          },
+        ],
+      },
+    });
+
+    /** Open block tunes menu */
+    cy.get('[data-cy=editorjs]')
+      .get('.cdx-block')
+      .click();
+
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-toolbar__settings-btn')
+      .click();
+
+    /** Check item displayed */
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-popover__container')
+      .get('[data-item-name="test-item"]')
+      .should('be.visible');
+
+    /** Check separator displayed */
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-popover__container')
+      .get('.ce-popover-item-separator')
+      .should('be.visible');
+  });
+
+  it('should perform keyboard navigation between items ignoring separators', () => {
+    /** Tool class to test how it is displayed inside block tunes popover */
+    class TestTune {
+      public static isTune = true;
+
+      /** Tool data displayed in block tunes popover */
+      public render(): MenuConfig {
+        return  [
+          {
+            onActivate: (): void => {},
+            icon: 'Icon',
+            title: 'Tune 1',
+            name: 'test-item-1',
+          },
+          {
+            type: PopoverItemType.Separator,
+          },
+          {
+            onActivate: (): void => {},
+            icon: 'Icon',
+            title: 'Tune 2',
+            name: 'test-item-2',
+          },
+        ];
+      }
+    }
+
+    /** Create editor instance */
+    cy.createEditor({
+      tools: {
+        testTool: TestTune,
+      },
+      tunes: [ 'testTool' ],
+      data: {
+        blocks: [
+          {
+            type: 'paragraph',
+            data: {
+              text: 'Hello',
+            },
+          },
+        ],
+      },
+    });
+
+    /** Open block tunes menu */
+    cy.get('[data-cy=editorjs]')
+      .get('.cdx-block')
+      .click();
+
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-toolbar__settings-btn')
+      .click();
+
+    /** Press Tab */
+    // eslint-disable-next-line cypress/require-data-selectors -- cy.tab() not working here
+    cy.get('body').tab();
+
+    /** Check first item is focused */
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-popover__container')
+      .get('[data-item-name="test-item-1"].ce-popover-item--focused')
+      .should('exist');
+
+    /** Check second item is not focused */
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-popover__container')
+      .get('[data-item-name="test-item-2"].ce-popover-item--focused')
+      .should('not.exist');
+
+    /** Press Tab */
+    // eslint-disable-next-line cypress/require-data-selectors -- cy.tab() not working here
+    cy.get('body').tab();
+
+    /** Check first item is not focused */
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-popover__container')
+      .get('[data-item-name="test-item-1"].ce-popover-item--focused')
+      .should('not.exist');
+
+    /** Check second item is focused */
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-popover__container')
+      .get('[data-item-name="test-item-2"].ce-popover-item--focused')
+      .should('exist');
+  });
+
+  it('should perform keyboard navigation between items ignoring separators when search query is applied', () => {
+    /** Tool class to test how it is displayed inside block tunes popover */
+    class TestTune {
+      public static isTune = true;
+
+      /** Tool data displayed in block tunes popover */
+      public render(): MenuConfig {
+        return  [
+          {
+            onActivate: (): void => {},
+            icon: 'Icon',
+            title: 'Tune 1',
+            name: 'test-item-1',
+          },
+          {
+            type: PopoverItemType.Separator,
+          },
+          {
+            onActivate: (): void => {},
+            icon: 'Icon',
+            title: 'Tune 2',
+            name: 'test-item-2',
+          },
+        ];
+      }
+    }
+
+    /** Create editor instance */
+    cy.createEditor({
+      tools: {
+        testTool: TestTune,
+      },
+      tunes: [ 'testTool' ],
+      data: {
+        blocks: [
+          {
+            type: 'paragraph',
+            data: {
+              text: 'Hello',
+            },
+          },
+        ],
+      },
+    });
+
+    /** Open block tunes menu */
+    cy.get('[data-cy=editorjs]')
+      .get('.cdx-block')
+      .click();
+
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-toolbar__settings-btn')
+      .click();
+
+    /** Check separator displayed */
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-popover__container')
+      .get('.ce-popover-item-separator')
+      .should('be.visible');
+
+    /** Enter search query */
+    cy.get('[data-cy=editorjs]')
+      .get('[data-cy=block-tunes] .cdx-search-field__input')
+      .type('Tune');
+
+    /** Check separator not displayed */
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-popover__container')
+      .get('.ce-popover-item-separator')
+      .should('not.be.visible');
+
+    /** Press Tab */
+    // eslint-disable-next-line cypress/require-data-selectors -- cy.tab() not working here
+    cy.get('body').tab();
+
+    /** Check first item is focused */
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-popover__container')
+      .get('[data-item-name="test-item-1"].ce-popover-item--focused')
+      .should('exist');
+
+    /** Check second item is not focused */
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-popover__container')
+      .get('[data-item-name="test-item-2"].ce-popover-item--focused')
+      .should('not.exist');
+
+    /** Press Tab */
+    // eslint-disable-next-line cypress/require-data-selectors -- cy.tab() not working here
+    cy.get('body').tab();
+
+    /** Check first item is not focused */
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-popover__container')
+      .get('[data-item-name="test-item-1"].ce-popover-item--focused')
+      .should('not.exist');
+
+    /** Check second item is focused */
+    cy.get('[data-cy=editorjs]')
+      .get('.ce-popover__container')
+      .get('[data-item-name="test-item-2"].ce-popover-item--focused')
+      .should('exist');
+  });
+
+  describe('Inline Popover', () => {
+    it('should open nested popover on click instead of hover', () => {
+      cy.createEditor({
+        tools: {
+          header: {
+            class: Header,
+          },
+        },
+        data: {
+          blocks: [
+            {
+              type: 'paragraph',
+              data: {
+                text: 'First block text',
+              },
+            },
+          ],
+        },
+      });
+
+      /** Open Inline Toolbar */
+      cy.get('[data-cy=editorjs]')
+        .find('.ce-paragraph')
+        .selectText('block');
+
+      /** Hover Convert To item which has nested popover */
+      cy.get('[data-cy=editorjs]')
+        .get('[data-item-name=convert-to]')
+        .trigger('mouseover');
+
+      /** Check nested popover didn't open */
+      cy.get('[data-cy=editorjs]')
+        .get('.ce-popover--nested .ce-popover__container')
+        .should('not.exist');
+
+      /** Click Convert To item which has nested popover */
+      cy.get('[data-cy=editorjs]')
+        .get('[data-item-name=convert-to]')
+        .click();
+
+      /** Check nested popover opened */
+      cy.get('[data-cy=editorjs]')
+        .get('.ce-popover--nested .ce-popover__container')
+        .should('exist');
+    });
+
+    it('should support keyboard nevigation between items', () => {
+      cy.createEditor({
+        tools: {
+          header: {
+            class: Header,
+          },
+        },
+        data: {
+          blocks: [
+            {
+              type: 'paragraph',
+              data: {
+                text: 'First block text',
+              },
+            },
+          ],
+        },
+      });
+
+      /** Open Inline Toolbar */
+      cy.get('[data-cy=editorjs]')
+        .find('.ce-paragraph')
+        .selectText('block');
+
+      /** Check Inline Popover opened */
+      cy.get('[data-cy=editorjs]')
+        .get('.ce-inline-toolbar .ce-popover__container')
+        .should('be.visible');
+
+      /** Check first item is NOT focused */
+      cy.get('[data-cy=editorjs]')
+        .get('.ce-inline-toolbar .ce-popover__container')
+        .get('[data-item-name="convert-to"].ce-popover-item--focused')
+        .should('not.exist');
+
+      /** Press Tab */
+      cy.tab();
+
+      /** Check first item became focused after tab */
+      cy.get('[data-cy=editorjs]')
+        .get('.ce-inline-toolbar .ce-popover__container')
+        .get('[data-item-name="convert-to"].ce-popover-item--focused')
+        .should('exist');
+
+      /** Check second item is NOT focused */
+      cy.get('[data-cy=editorjs]')
+        .get('.ce-inline-toolbar .ce-popover__container')
+        .get('[data-item-name="link"] .ce-popover-item--focused')
+        .should('not.exist');
+
+      /** Press Tab */
+      cy.tab();
+
+      /** Check second item became focused after tab */
+      cy.get('[data-cy=editorjs]')
+        .get('.ce-inline-toolbar .ce-popover__container')
+        .get('[data-item-name="link"] .ce-popover-item--focused')
+        .should('exist');
+    });
+
+    it.only('should allow to reach nested popover via keyboard', () => {
+      cy.createEditor({
+        tools: {
+          header: {
+            class: Header,
+          },
+        },
+        data: {
+          blocks: [
+            {
+              type: 'paragraph',
+              data: {
+                text: 'First block text',
+              },
+            },
+          ],
+        },
+      });
+
+      /** Open Inline Toolbar */
+      cy.get('[data-cy=editorjs]')
+        .find('.ce-paragraph')
+        .selectText('block');
+
+      /** Check Inline Popover opened */
+      cy.get('[data-cy=editorjs]')
+        .get('.ce-inline-toolbar .ce-popover__container')
+        .should('be.visible');
+
+      /** Press Tab */
+      cy.tab();
+
+      /** Press Tab */
+      cy.get('[data-item-name="convert-to"]')
+        .type('{enter}');
+
+      /** Check Inline Popover opened */
+      cy.get('[data-cy=editorjs]')
+        .get('.ce-inline-toolbar .ce-popover--nested .ce-popover__container')
+        .should('be.visible');
+
+      /** Check first item is NOT focused */
+      cy.get('[data-cy=editorjs]')
+        .get('.ce-popover__container')
+        .get('[data-item-name="header"].ce-popover-item--focused')
+        .should('not.exist');
+
+      /** Press Tab */
+      // eslint-disable-next-line cypress/require-data-selectors -- cy.tab() not working here
+      cy.get('body').tab();
+
+      /** Check first item is focused */
+      cy.get('[data-cy=editorjs]')
+        .get('.ce-popover__container')
+        .get('[data-item-name="header"].ce-popover-item--focused')
+        .should('exist');
+    });
   });
 });